Allow test methods with no return type hint
Summary: A lot of test files are not in strict mode, and thereby their test methods do not have return type hints. This makes filtering test methods hard. In these cases, accept a method as a test method if it begins with 'test' and is public. This fixes DenyIfInvalidCampaignPredictionTransitionRuleTest and more.

Reviewed By: kmeht, fredemmott

Differential Revision: D9281538

fbshipit-source-id: 264a35eb7b1c6c5594c062037eb36f02f794a08f