Deal with hacky data provider
Summary:
AdEntityValidatorTestTrait has test methods that are not meant to be directly called if the overriden `dataProviderInputs` returns an empty container, because the test methods in the original class invoke them.

```
// hack to include AdEntityValidatorTestTrait so that genErrors() can
// call the testGenValidateFoo() methods
<<__Override>>
public function dataProviderInputs(
): darray<string, (AdDeliveryValidator, EntAd, varray<int>)> {
  return darray[];
}
```

But when the data provider returns non-empty data, the test methods in the trait need to be executed.

This should fix all of the tests in flib/ads using this trait; hopefully, there are no other hacky data providers like this in the codebase.

Reviewed By: kmeht

Differential Revision: D9324173

fbshipit-source-id: 35605f53c2b0ed6b982740b3c497cf5077587d13